+++ forge/src/main/java/com/yanny/ali/plugin/mods/farmers_delight_glm/Plugin.java
@@ -1,0 +1,20 @@
+package com.yanny.ali.plugin.mods.farmers_delight_glm;
+
+import com.yanny.ali.api.AliEntrypoint;
+import com.yanny.ali.plugin.IForgePlugin;
+
+@AliEntrypoint
+public class Plugin implements IForgePlugin {
+    @Override
+    public String getModId() {
+        return "farmersdelight";
+    }
+
+    @Override
+    public void registerGLM(IRegistry registry) {
+//        GlobalLootModifierUtils.registerGlobalLootModifier(registry, AddItemModifier.class);
+//        GlobalLootModifierUtils.registerGlobalLootModifier(registry, AddLootTableModifier.class);
+//        GlobalLootModifierUtils.registerGlobalLootModifier(registry, PastrySlicingModifier.class);
+//        GlobalLootModifierUtils.registerGlobalLootModifier(registry, ReplaceItemModifier.class);
+    }
+}