+++ common/src/main/java/com/yanny/ali/compatibility/JeiCompatibility.java
@@ -25,6 +25,7 @@ import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 
 import java.util.*;
+import java.util.concurrent.CancellationException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeoutException;
@@ -83,17 +84,19 @@ public class JeiCompatibility implements IModPlugin {
             byte[] fullCompressedData = futureData.get();
 
             registerData(registration, fullCompressedData);
+            LOGGER.info("Data registration finished successfully.");
         } catch (ExecutionException e) {
             Throwable cause = e.getCause() != null ? e.getCause() : e;
 
-            if (cause instanceof TimeoutException) {
-                futureData.cancel(true);
-                PluginManager.CLIENT_REGISTRY.clearLootData();
-                LOGGER.error("Failed to received data: Inactivity timeout occurred. Registration aborted!");
+            if (cause instanceof TimeoutException || cause instanceof CancellationException) {
+                LOGGER.error("Failed to receive data: Operation aborted or timed out. Registration aborted!");
             } else {
                 LOGGER.error("Failed to finish registering data with error {}", cause.getMessage());
                 cause.printStackTrace();
             }
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            LOGGER.error("Registration thread interrupted!");
         } catch (Throwable e) {
             e.printStackTrace();
             LOGGER.error("Failed to finish registering data with unexpected error {}", e.getMessage());