+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# []
+
+- Added GLM support
+- Fixed missing spawn eggs for modded mobs on (Neo)Forge
+
 ## [1.5.2]
 
 - Fixed missing Wandering Trader trade info