+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added GLM support
 - Fixed missing spawn eggs for modded mobs on (Neo)Forge
+- Hiding injected loot tables
 
 ## [1.5.2]