+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Hiding injected loot tables
 - Improved loot info for unknown entries/modifiers/predicates
 - Displaying loot table name in gameplay tooltip
+- Fixed occasional no loot information
 
 ## [1.5.2]